requirements.txt
async-generator==1.10
attrs==22.1.0
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
h11==0.13.0
idna==3.3
outcome==1.2.0
pycodestyle==2.9.1
PyJWT==2.4.0
PySocks==1.7.1
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.2.1
pywatchman==1.4.1
requests==2.28.1
selenium==4.4.3
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.21.0
trio-websocket==0.9.2
twilio==7.13.0
undetected-chromedriver==3.1.5.post4
urllib3==1.26.12
websockets==10.3
wsproto==1.2.0
